s3ext: refactor s3wrapper and s3downloader codes
refactor, optimize and rename them to gpreader and s3reader.
