+++ gpAux/extensions/gps3ext/Makefile.check
@@ -16,7 +16,7 @@ endif
 
 # Targets
 PROGRAM = gpcheckcloud
-OBJS = src/gpcheckcloud.o src/s3conf.o src/s3downloader.o src/s3wrapper.o src/s3utils.o src/s3log.o src/s3common.o lib/http_parser.o lib/ini.o src/s3url_parser.o src/s3http_headers.o src/s3thread.o
+OBJS = src/gpcheckcloud.o src/s3conf.o src/gpreader.o src/s3utils.o src/s3log.o src/s3common.o lib/http_parser.o lib/ini.o src/s3url_parser.o src/s3http_headers.o src/s3thread.o src/s3extbase.o src/s3reader.o
 
 # Launch
 PGXS := $(shell pg_config --pgxs)
