+++ gpAux/extensions/gps3ext/include/gpcheckcloud.h
@@ -7,9 +7,10 @@
 
 #include "s3common.h"
 #include "s3conf.h"
-#include "s3downloader.h"
+#include "s3reader.h"
 #include "s3log.h"
-#include "s3wrapper.h"
+#include "gpreader.h"
+#include "s3thread.h"
 
 #define BUF_SIZE 64 * 1024
