+++ gpAux/extensions/gps3ext/include/gps3ext.h
@@ -1,5 +1,5 @@
-#ifndef _GPS3EXT_H_
-#define _GPS3EXT_H_
+#ifndef __GP_S3EXT_H_
+#define __GP_S3EXT_H_
 #include <string>
 
 #include <signal.h>
@@ -15,8 +15,8 @@ typedef struct GpId {
     int4 numsegments; /* count of distinct segindexes */
     int4 dbid;        /* the dbid of this database */
     int4 segindex;    /* content indicator: -1 for entry database,
-                       * 0, ..., n-1 for segment database *
-                       * a primary and its mirror have the same segIndex */
+                    * 0, ..., n-1 for segment database *
+                    * a primary and its mirror have the same segIndex */
 } GpId;
 extern GpId GpIdentity;
