+++ gpAux/extensions/gps3ext/include/s3common.h
@@ -19,7 +19,13 @@ struct S3Credential {
     string secret;
 };
 
-enum Method { GET, PUT, POST, DELETE, HEAD };
+enum Method {
+    GET,
+    PUT,
+    POST,
+    DELETE,
+    HEAD
+};
 
 bool SignRequestV4(const string& method, HTTPHeaders* h,
                    const string& orig_region, const string& path,
