+++ gpAux/extensions/gps3ext/include/s3reader.h
@@ -1,5 +1,5 @@
-#ifndef __S3DOWNLOADER_H__
-#define __S3DOWNLOADER_H__
+#ifndef __S3_READER_H__
+#define __S3_READER_H__
 
 #include <fcntl.h>
 #include <pthread.h>
@@ -37,11 +37,17 @@ typedef enum compression_type {
 class OffsetMgr {
    public:
     OffsetMgr(uint64_t maxsize, uint64_t chunksize);
-    ~OffsetMgr() { pthread_mutex_destroy(&this->offset_lock); };
+    ~OffsetMgr() {
+        pthread_mutex_destroy(&this->offset_lock);
+    };
     Range NextOffset();  // ret.len == 0 means EOF
     void Reset(uint64_t n);
-    uint64_t Chunksize() { return this->chunksize; };
-    uint64_t Size() { return this->maxsize; };
+    uint64_t Chunksize() {
+        return this->chunksize;
+    };
+    uint64_t Size() {
+        return this->maxsize;
+    };
 
    private:
     pthread_mutex_t offset_lock;
@@ -57,8 +63,12 @@ class BlockingBuffer {
     BlockingBuffer(const string& url, OffsetMgr* o);
     virtual ~BlockingBuffer();
     bool Init();
-    bool EndOfFile() { return this->eof; };
-    bool Error() { return this->error; };
+    bool EndOfFile() {
+        return this->eof;
+    };
+    bool Error() {
+        return this->error;
+    };
 
     uint64_t Read(char* buf, uint64_t len);
     uint64_t Fill();
@@ -138,7 +148,9 @@ class HTTPFetcher : public BlockingBuffer {
 
    protected:
     uint64_t fetchdata(uint64_t offset, char* data, uint64_t len);
-    virtual bool processheader() { return true; };
+    virtual bool processheader() {
+        return true;
+    };
     CURL* curl;
     Method method;
     HTTPHeaders headers;
@@ -149,7 +161,7 @@ class S3Fetcher : public HTTPFetcher {
    public:
     S3Fetcher(const string& url, const string& region, OffsetMgr* o,
               const S3Credential& cred);
-    ~S3Fetcher(){};
+    ~S3Fetcher() {};
 
    protected:
     virtual bool processheader();
@@ -177,8 +189,12 @@ struct BucketContent {
                                                   uint64_t size);
     BucketContent();
     ~BucketContent();
-    string Key() const { return this->key; };
-    uint64_t Size() const { return this->size; };
+    string Key() const {
+        return this->key;
+    };
+    uint64_t Size() const {
+        return this->size;
+    };
 
    private:
     // BucketContent(const BucketContent& b) = delete;
