+++ gpAux/extensions/gps3ext/include/s3utils.h
@@ -1,10 +1,10 @@
 #ifndef __S3_UTILFUNCTIONS__
 #define __S3_UTILFUNCTIONS__
 
-#include <cstdio>
-#include <cstdlib>
 #include <stdint.h>
 #include <sys/types.h>
+#include <cstdio>
+#include <cstdlib>
 #include <cstring>
 #include <string>
 
@@ -44,7 +44,7 @@ size_t find_Nth(const string& str,  // where to work
 class MD5Calc {
    public:
     MD5Calc();
-    ~MD5Calc(){};
+    ~MD5Calc() {};
     bool Update(const char* data, int len);
     const char* Get();
 
@@ -58,13 +58,23 @@ class DataBuffer {
    public:
     DataBuffer(uint64_t size);
     ~DataBuffer();
-    void reset() { length = 0; };
+    void reset() {
+        length = 0;
+    };
 
     uint64_t append(const char* buf, uint64_t len);  // ret < len means full
-    const char* getdata() { return data; };
-    uint64_t len() { return this->length; };
-    bool full() { return maxsize == length; };
-    bool empty() { return 0 == length; };
+    const char* getdata() {
+        return data;
+    };
+    uint64_t len() {
+        return this->length;
+    };
+    bool full() {
+        return maxsize == length;
+    };
+    bool empty() {
+        return 0 == length;
+    };
 
    private:
     const uint64_t maxsize;
@@ -81,7 +91,9 @@ class Config {
                const string& defaultvalue);
     bool Scan(const string& sec, const string& key, const char* scanfmt,
               void* dst);
-    void* Handle() { return (void*)this->_conf; };
+    void* Handle() {
+        return (void*)this->_conf;
+    };
 
    private:
     ini_t* _conf;
