+++ gpAux/extensions/gps3ext/src/s3thread.cpp
@@ -1,5 +1,5 @@
-#include <pthread.h>
 #include <openssl/crypto.h>
+#include <pthread.h>
 
 #define MUTEX_TYPE pthread_mutex_t
 #define MUTEX_SETUP(x) pthread_mutex_init(&(x), NULL)
