+++ gpAux/extensions/gps3ext/test/s3reader_test.cpp
@@ -1,4 +1,4 @@
-#include "s3downloader.cpp"
+#include "s3reader.cpp"
 #include "gtest/gtest.h"
 
 volatile bool QueryCancelPending = false;
